Opinionated settings UI rework
- New "Travel & Pathfinding" settings tab with mount, pathfinding, stuck detection, and crafting return spot settings
- Option to avoid stellar return for pathing (still used for vendor trips)
- Renamed moon sprint to stellar sprint
- Mission name now opens a detail popup window, clicking again brings it to front - doubles as a nifty stats monitor
- Removed redundant "All Settings" page and "Show/Hide Tabs" section
- Horizontal scrollbar on mission tables and credit shopping when window is narrow
- Misc settings now uses a cleaner linear layout
- Added tooltips for score per hour/minute toggle and stellar status removal
