+++ ICE/ConfigFiles/Config_Misc.cs
@@ -20,6 +20,7 @@ public partial class Config
     public bool StartUponEnterMoon { get; set; } = false;
     public bool PersonalReturnSpot { get; set; } = false;
     public bool ClosestNodeSelection { get; set; } = false;
+    public bool AvoidStellarReturn { get; set; } = false;
     public bool RandomizeWaypoints { get; set; } = false;
     public float RandomizeWaypointsRadius { get; set; } = 1.0f;
     public bool RandomizeWaypointsDebug { get; set; } = false;
