+++ ICE/Ui/MainUi/Settings/Settings_Table/TravelSettings.cs
@@ -1,0 +1,343 @@
+using Dalamud.Interface;
+using ECommons.GameHelpers;
+using FFXIVClientStructs.FFXIV.Client.Game.UI;
+using Lumina.Excel.Sheets;
+using Pictomancy;
+using System.Collections.Generic;
+
+namespace ICE.Ui.MainUi.Settings.Settings_Table
+{
+    internal class TravelSettings
+    {
+        private static bool visualizeRadius = false;
+        private static bool visualizeDismountRadius = false;
+        private static Dictionary<uint, string> availableMounts = new();
+
+        private static string mountSearchText = "";
+        private static int mountDisplayOffset = 0;
+        private static int mountItemsPerPage = 10;
+
+        public static unsafe void Draw()
+        {
+            MountSelection();
+            ImGui.Dummy(new Vector2(0, 5));
+            ImGui.Separator();
+            ImGui.Dummy(new Vector2(0, 5));
+            PathfindingSettings();
+            ImGui.Dummy(new Vector2(0, 5));
+            ImGui.Separator();
+            ImGui.Dummy(new Vector2(0, 5));
+            StuckSettings();
+            ImGui.Dummy(new Vector2(0, 5));
+            ImGui.Separator();
+            ImGui.Dummy(new Vector2(0, 5));
+            CraftingLocations();
+        }
+
+        private static unsafe void MountSelection()
+        {
+            ImGuiEx.IconWithText(FontAwesomeIcon.Feather, "Mount Settings");
+            ImGui.Dummy(new Vector2(0, 5));
+
+            bool mountOutsideMission = C.UseMountOutsideMission;
+            bool mountInMission = C.UseMountInMission;
+            float minMountRange = C.MountRadius;
+            float dismountRange = C.DismountRadius;
+
+            if (ImGui.Button("Select Mounting Option"))
+            {
+                availableMounts.Clear();
+                availableMounts[0] = "Mount Roulette";
+
+                var mountSheet = Svc.Data.GetExcelSheet<Mount>();
+
+                foreach (var mountItem in mountSheet)
+                {
+                    if (!PlayerState.Instance()->IsMountUnlocked(mountItem.RowId)) continue;
+
+                    string mountName = System.Globalization.CultureInfo.CurrentCulture.TextInfo.ToTitleCase(mountItem.Singular.ToString().ToLower());
+                    uint id = mountItem.RowId;
+
+                    availableMounts[id] = mountName;
+                }
+
+                mountSearchText = "";
+                mountDisplayOffset = 0;
+
+                ImGui.OpenPopup("Mount Options");
+            }
+            ImGui.SameLine();
+            ImGui.AlignTextToFramePadding();
+            ImGui.Text($"Mount: {C.MountName}");
+
+            if (ImGui.BeginPopup("Mount Options"))
+            {
+                ImGui.InputText("Search", ref mountSearchText, 100);
+
+                var filteredMounts = availableMounts
+                    .Where(kvp => string.IsNullOrEmpty(mountSearchText) ||
+                                  kvp.Value.Contains(mountSearchText, StringComparison.OrdinalIgnoreCase))
+                    .ToList();
+
+                int totalItems = filteredMounts.Count;
+                int maxOffset = Math.Max(0, totalItems - mountItemsPerPage);
+                mountDisplayOffset = Math.Min(mountDisplayOffset, maxOffset);
+
+                var displayMounts = filteredMounts
+                    .Skip(mountDisplayOffset)
+                    .Take(mountItemsPerPage);
+
+                foreach (var mount in displayMounts)
+                {
+                    if (ImGui.Selectable($"{mount.Value}##{mount.Key}"))
+                    {
+                        C.MountId = mount.Key;
+                        C.MountName = mount.Value;
+                        C.Save();
+                        ImGui.CloseCurrentPopup();
+                    }
+                }
+
+                ImGui.Separator();
+
+                if (ImGui.Button("Previous") && mountDisplayOffset > 0)
+                {
+                    mountDisplayOffset = Math.Max(0, mountDisplayOffset - mountItemsPerPage);
+                }
+
+                ImGui.SameLine();
+                ImGui.Text($"{mountDisplayOffset + 1}-{Math.Min(mountDisplayOffset + mountItemsPerPage, totalItems)} of {totalItems}");
+
+                ImGui.SameLine();
+                if (ImGui.Button("Next") && mountDisplayOffset < maxOffset)
+                {
+                    mountDisplayOffset = Math.Min(maxOffset, mountDisplayOffset + mountItemsPerPage);
+                }
+
+                ImGui.EndPopup();
+            }
+
+            if (ImGui.Checkbox("Use mount outside mission", ref mountOutsideMission))
+            {
+                C.UseMountOutsideMission = mountOutsideMission;
+                C.Save();
+            }
+
+            if (ImGui.Checkbox("Use mount in mission", ref mountInMission))
+            {
+                C.UseMountInMission = mountInMission;
+                C.Save();
+            }
+
+            ImGui.SetNextItemWidth(100);
+            if (ImGui.DragFloat("Minimum Mounting Range", ref minMountRange, 1))
+            {
+                C.MountRadius = minMountRange;
+                C.Save();
+            }
+            ImGui.Checkbox("Visualize radius", ref visualizeRadius);
+            ImGui.SetNextItemWidth(100);
+            if (ImGui.DragFloat("Dismount Target Range", ref dismountRange, 1))
+            {
+                C.DismountRadius = dismountRange;
+                C.Save();
+            }
+            ImGui.Checkbox("Visualize Dismount Radius", ref visualizeDismountRadius);
+
+            using (var drawList = PictoService.Draw(hints: Utils.GetPictoHints()))
+            {
+                if (drawList == null)
+                    return;
+
+                var playerPos = Player.Position;
+
+                if (visualizeRadius)
+                    PictoService.VfxRenderer.AddCircle("Mount_Radius Circle", playerPos, C.MountRadius, Utils.FromUintABGR(2616716297));
+                if (visualizeDismountRadius)
+                    PictoService.VfxRenderer.AddCircle("Dismount_Radius Circle", playerPos, C.DismountRadius, Utils.FromUintABGR(2601121571));
+            }
+        }
+
+        private static void PathfindingSettings()
+        {
+            ImGuiEx.IconWithText(FontAwesomeIcon.Route, "Pathfinding");
+            ImGui.Dummy(new Vector2(0, 5));
+
+            bool stellarSprint = C.MoonSprint;
+            if (ImGui.Checkbox("Auto-Use Stellar Sprint", ref stellarSprint))
+            {
+                C.MoonSprint = stellarSprint;
+                C.Save();
+            }
+
+            bool closestNode = C.ClosestNodeSelection;
+            if (ImGui.Checkbox("Prioritize closest gathering node", ref closestNode))
+            {
+                C.ClosestNodeSelection = closestNode;
+                C.Save();
+            }
+            if (ImGui.IsItemHovered())
+            {
+                ImGui.SetTooltip("Always navigate to the closest targetable node instead of following the fixed route order.\nUseful for timed EX+ missions where speed matters.");
+            }
+
+            bool randomize = C.RandomizeWaypoints;
+            if (ImGui.Checkbox("Randomize waypoint positions", ref randomize))
+            {
+                C.RandomizeWaypoints = randomize;
+                C.Save();
+            }
+            if (ImGui.IsItemHovered())
+            {
+                ImGui.SetTooltip("Adds a small random offset to navigation destinations so the character doesn't always follow the exact same path");
+            }
+            if (randomize)
+            {
+                float radius = C.RandomizeWaypointsRadius;
+                ImGui.SetNextItemWidth(100);
+                if (ImGui.SliderFloat("Randomize radius (yalms)", ref radius, 0.5f, 1.0f, "%.1f"))
+                {
+                    C.RandomizeWaypointsRadius = radius;
+                    C.SaveDebounced();
+                }
+                bool showDebug = C.RandomizeWaypointsDebug;
+                if (ImGui.Checkbox("Show random location debug target", ref showDebug))
+                {
+                    C.RandomizeWaypointsDebug = showDebug;
+                    C.Save();
+                }
+            }
+
+            bool avoidStellarReturn = C.AvoidStellarReturn;
+            if (ImGui.Checkbox("Avoid Stellar Return for pathing", ref avoidStellarReturn))
+            {
+                C.AvoidStellarReturn = avoidStellarReturn;
+                C.Save();
+            }
+            if (ImGui.IsItemHovered())
+            {
+                ImGui.SetTooltip("When enabled, the pathfinder will never use Stellar Return as a shortcut.\nIt will still be used when returning to hub for vendors/repairs.");
+            }
+
+            bool useHubReturn = C.UseHubReturn;
+            if (ImGui.Checkbox("Use Hub Return", ref useHubReturn))
+            {
+                C.UseHubReturn = useHubReturn;
+                C.Save();
+            }
+
+            bool useAethernet = C.UseAethernet;
+            if (ImGui.Checkbox("Use Aethernet", ref useAethernet))
+            {
+                C.UseAethernet = useAethernet;
+                C.Save();
+            }
+
+            var minHubReturnDistance = C.HubReturn_Distance;
+            ImGui.SetNextItemWidth(200);
+            if (ImGui.DragFloat("Distance before hub return is used", ref minHubReturnDistance))
+            {
+                C.HubReturn_Distance = minHubReturnDistance;
+                C.SaveDebounced();
+            }
+
+            bool DisableRedAlertPathing = C.DisablePathfindingToRedAlert;
+            if (ImGui.Checkbox("Disable Pathfinding to Red Alerts", ref DisableRedAlertPathing))
+            {
+                C.DisablePathfindingToRedAlert = DisableRedAlertPathing;
+                C.Save();
+            }
+        }
+
+        private static void StuckSettings()
+        {
+            ImGuiEx.IconWithText(FontAwesomeIcon.ExclamationTriangle, "Stuck Detection");
+            ImGui.Dummy(new Vector2(0, 5));
+
+            bool unstuckEnabled = C.JumpIfStuck_V2 || C.RetargetIfStuck;
+            if (ImGui.Checkbox("If stuck during nav movement:", ref unstuckEnabled))
+            {
+                if (unstuckEnabled)
+                    C.JumpIfStuck_V2 = true;
+                else
+                {
+                    C.JumpIfStuck_V2 = false;
+                    C.RetargetIfStuck = false;
+                }
+                C.Save();
+            }
+            ImGui.SameLine();
+            ImGuiEx.HelpMarker(
+                "When stuck during navmesh movement for the configured delay:\n" +
+                "- Jump: attempts to jump over the obstacle\n" +
+                "- Retarget: stops and re-pathfinds to the destination (re-randomizes if enabled)");
+            if (!unstuckEnabled) ImGui.BeginDisabled();
+            if (ImGui.RadioButton("Jump", C.JumpIfStuck_V2 && !C.RetargetIfStuck))
+            {
+                C.JumpIfStuck_V2 = true;
+                C.RetargetIfStuck = false;
+                C.Save();
+            }
+            ImGui.SameLine();
+            if (ImGui.RadioButton("Retarget", C.RetargetIfStuck))
+            {
+                C.RetargetIfStuck = true;
+                C.JumpIfStuck_V2 = false;
+                C.Save();
+            }
+            ImGui.SameLine();
+            ImGui.Text("after");
+            ImGui.SameLine();
+            int stuckDelay = C.StuckDelayMs;
+            ImGui.SetNextItemWidth(100);
+            if (ImGui.SliderInt("ms stuck###StuckDelay", ref stuckDelay, 500, 3000))
+            {
+                if (C.StuckDelayMs != stuckDelay)
+                {
+                    C.StuckDelayMs = stuckDelay;
+                    C.SaveDebounced();
+                }
+            }
+            if (!unstuckEnabled) ImGui.EndDisabled();
+        }
+
+        private static void CraftingLocations()
+        {
+            ImGuiEx.IconWithText(FontAwesomeIcon.MapPin, "Crafting Return Spot");
+            ImGui.Dummy(new Vector2(0, 5));
+
+            bool usePersonalLocations = C.PersonalReturnSpot;
+            if (ImGui.Checkbox("Use personal return spots", ref usePersonalLocations))
+            {
+                C.PersonalReturnSpot = usePersonalLocations;
+                C.Save();
+            }
+            if (usePersonalLocations)
+            {
+                var territory = Player.Territory.RowId;
+                var location = Player.Position;
+                ImGui.SameLine();
+                if (C.CrafterLocations.TryGetValue(territory, out var moonLoc))
+                {
+                    if (ImGui.Button("Set to current location"))
+                    {
+                        C.CrafterLocations[territory] = location;
+                        C.Save();
+                    }
+                    ImGui.SameLine();
+                    ImGui.Text($"({moonLoc.X:N1}, {moonLoc.Y:N1}, {moonLoc.Z:N1})");
+                }
+                else
+                {
+                    if (ImGui.Button("Add Location"))
+                    {
+                        C.CrafterLocations[territory] = Player.Position;
+                        C.Save();
+                    }
+                    ImGui.SameLine();
+                    ImGui.Text("No location set");
+                }
+            }
+        }
+    }
+}
